# Copy object.di from object_.d