--- a/src/object_.d
+++ b/src/object_.d
@@ -150,16 +150,7 @@ class Object
     }
 }
 
-/************************
- * Returns true if lhs and rhs are equal.
- */
-bool opEquals(const Object lhs, const Object rhs)
-{
-    // A hack for the moment.
-    return opEquals(cast()lhs, cast()rhs);
-}
-
-bool opEquals(Object lhs, Object rhs)
+auto opEquals(Object lhs, Object rhs)
 {
     // If aliased to the same object or both null => equal
     if (lhs is rhs) return true;
@@ -179,6 +170,15 @@ bool opEquals(Object lhs, Object rhs)
 
     // General case => symmetric calls to method opEquals
     return lhs.opEquals(rhs) && rhs.opEquals(lhs);
+}
+
+/************************
+* Returns true if lhs and rhs are equal.
+*/
+auto opEquals(const Object lhs, const Object rhs)
+{
+    // A hack for the moment.
+    return opEquals(cast()lhs, cast()rhs);
 }
 
 /**