Updating object.di and object_.d is a maintenance liability. This PR removes non-public code from object._d and adds a bit from object.di in order to generate the latter automatically.

Comparing object.di to the generated header reveals:
- aliases for string,wstring,dstring not converted correctly, needs https://github.com/D-Programming-Language/dmd/pull/4594
- overriding methods often not declared in classes declared in object.di. This can cause unexpected results when deriving from this class and calling `super.method()`.
- different types for `TpeInfo_Class.defaultConstructor`
- different return types for `TypeInfo_Struct.xtoHash`
- missing fields in classes `MemberInfo_field` and `MemberInfo_function`, can have bad effects when deriving from it or using `__traits(classInstanceSize)`.
- different next/base field declarations for `TypeInfo_Const`. See also https://github.com/D-Programming-Language/druntime/pull/1182
- header generation moves function attributes before the return type.

Unfortunately, neither the generated object_.di nor the original object_.d can be used as a replacement for object.di yet:
- object_.d does not work because it has (now local) imports to the rt package which are not supposed to be visible in general.
- object_.di does not contain all the source code necessary for CTFE. Adding `auto` can work in some places, but not for the `Exception` and `Error` constructors.
- in addition, the missing source in object_.di can disable some inlining opportunities, especially for the AA templates.

So we need better control over the omission of the code in the generated headers. Adding inline hints might be a possible solution, though somewhat abusive in the case of CTFE.