@MartinNowak 2015-04-18
I also wanted to look at that, though I thought about ditching `object.di` in favor of always using `object.d`. Any reason against that?