@rainers 2015-04-18
> I also wanted to look at that, though I thought about ditching object.di in favor of always using object.d. Any reason against that?

The problem is that the import references into the `rt` package are still analyzed, so you get errors in a non-druntime build. We would have to wrap these into extern(C) function or use `externDFunc`.