@rainers 2015-04-18
BTW: I think that `ModuleInfo` should not be in object.d, it's pretty internal to the runtime. The compiler expects it to be defined, though, but never uses the definition.