@MartinNowak 2015-04-18
> The problem is that the import references into the rt package are still analyzed, so you get errors in a non-druntime build. We would have to wrap these into extern(C) function or use externDFunc.

Then let's do that, the main reasons to import `rt` modules instead of using `extern (C)`, are also achieved using `externDFunc`.