@MartinNowak 2015-04-18
> BTW: I think that ModuleInfo should not be in object.d, it's pretty internal to the runtime.

You need to have the interface though, we could split it into an internal and public type though.