@rainers 2015-04-18
> Then let's do that, the main reasons to import rt modules instead of using extern (C), are also achieved using externDFunc.

Ok, will look into it. This will generate a dependency on core.internal.traits and core.demangle, though. With local imports that should be restricted to compiling with -inline.