@rainers 2015-04-18
> You need to have the interface though, we could split it into an internal and public type though.

I was rather thinking of moving it to another core module to import it when you actually need it.