@MartinNowak 2015-04-18
As long as `foreach (m; ModuleInfo)` keeps on working.