@rainers 2015-04-18
> Ok, will look into it.

Done. I had to modify rt.util.string a bit to not call templates through externDFunc.