@rainers 2015-04-18
> As long as foreach (m; ModuleInfo) keeps on working.

That would need a new import then. That change should be in another PR anyway.