@rainers 2015-04-19
See https://github.com/D-Programming-Language/dmd/pull/4597 for the testtypeid.d failure. I've no idea how to deal with the changes to the error outputs in the fail_compilation tests. Should I just create a failing PR to be merged at the same time?