@rainers 2015-04-19
FYI: Even though the size of object.di triples when being copied, I don't see a difference in compilation speed (slowness) when compiling a simple hello world like program using writeln. Both take slightly below 0.6 sec.