@MartinNowak 2015-04-19
> I don't see a difference in compilation speed (slowness) when compiling a simple hello world

I'll double check that. Your baseline shouldn't be hello world, which imports std.stdio, but an empty d file.