@MartinNowak 2015-04-19
> ../../druntime/import/object.di(1666): Error: uncaught CTFE exception object.Exception("abc"c)

The file and line number doesn't make sense, seems like it shows the location of the Exception constructor.