@MartinNowak 2015-04-19
Have you tried to directly use object.d?