@rainers 2015-04-19
> Your baseline shouldn't be hello world, which imports std.stdio, but an empty d file.

I thought there would be a little difference that I could weigh to a standard use case. But there was none.

> The file and line number doesn't make sense, seems like it shows the location of the Exception constructor.

There is a difference between object.di and object._d: object.di hides the implementation of the constructor of Throwable, so there is one entry less in the stack trace. I have no idea how this interacts with CTFE, though.

> Have you tried to directly use object.d?

I don't think there is a difference whether a .di or a .d file is used.