@MartinNowak 2015-04-19
> I don't think there is a difference whether a .di or a .d file is used.

Exactly, so we can just copy the .d file, and spare the header-gen troubles.

> There is a difference between object.di and object._d: object.di hides the implementation of the constructor of Throwable, so there is one entry less in the stack trace. I have no idea how this interacts with CTFE, though.

Still doesn't make sense, in `throw new Expr(arguments)` the stacktrace should start with `throw` not the constructor of the exception. But it seems like the CTFE code was programmed like that.