@MartinNowak 2015-04-19
[Issue 14465 – CTFE exception stacktrace shows location of Exception constructor](https://issues.dlang.org/show_bug.cgi?id=14465)
Anyhow, in the mean-time you might want to open a dmd pull that changes the expected test output.