@rainers 2015-04-19
> Issue 14465 – CTFE exception stacktrace shows location of Exception constructor
https://github.com/D-Programming-Language/dmd/pull/4599

> Anyhow, in the mean-time you might want to open a dmd pull that changes the expected test output.
https://github.com/D-Programming-Language/dmd/pull/4600