@rainers 2015-04-19
> Exactly, so we can just copy the .d file, and spare the header-gen troubles.

That's what the makefiles do in this PR now, although it copies to object.di to avoid troubles with stale files in peoples import folder.