@MartinNowak 2015-04-20
> That's what the makefiles do in this PR now, although it copies to object.di to avoid troubles with stale files in peoples import folder.

There is a semantic connotation in .di headers, if it's the plain source, we should name it .d.
You can temporarily add a `rm -f import/object.di` to the makefiles.