@MartinNowak 2015-04-20
Compile times for an empty .d file increase from 3.5ms to 6ms, compiling (not linking) a printf hello world goes from 7ms to 10ms. The increase comes from lexing overhead and is neglectable.
Though I do wonder what we could do to optimize the header, e.g. using -Hf or stripping out unittests.
There seems to be room for a libdparse based header gen.