@redstar 2015-05-01
I have not yet tried this PR with ldc but I really like it. The differences between `object_.d` and `object.di` caused some trouble.