@rainers 2015-05-01
> Compile times for an empty .d file increase from 3.5ms to 6ms, compiling (not linking) a printf hello world goes from 7ms to 10ms. The increase comes from lexing overhead and is neglectable.

I guess my benchmark batch uses the standard windows timer and has a resolution of 10ms only, so I did not notice the difference.

> Though I do wonder what we could do to optimize the header, e.g. using -Hf or stripping out unittests.
> There seems to be room for a libdparse based header gen.

Maybe some customization options of header generation to adopt to more use cases:
- keep all implementations (but still remove unittests)
- keep all implementations but those of private functions.
- expand version/debug conditionals