@rainers 2015-05-01
Failures caused by compiler generated identifiers appearing in error messages. See https://github.com/D-Programming-Language/dmd/pull/4624 for making the tests agnostic to these numbers.