@MartinNowak 2015-05-05
Auto-merge toggled on