@MartinNowak 2015-05-05
Thanks for this @rainers, it was on my todo list for too long already.