@andralex 2015-05-05
oh this is so awesome - I hated all that duplication