@MartinNowak 2015-05-06
Yikes! Forgot about that, @rainers can you check whether you converted to CRLF or so.