@rainers 2015-05-06
> Yikes! Forgot about that, @rainers can you check whether you converted to CRLF or so.

Arrgh, this creeps in somehow from time to time. I suspect it's caused by a bad .gitattributes in one of my branches, but I don't know how that could infect posix.mak.

https://github.com/D-Programming-Language/druntime/pull/1258 reverts the line endings to LF only. I guess there is no way to avoid polluting the history of posix.mak, isn't there?