@rainers 2015-05-06
I guess we should remove object.di from the source folder now? There is also https://github.com/D-Programming-Language/phobos/pull/3263 with a necessary change in phobos. I'm not sure why this isn't caught in any unittests.