@MartinNowak 2015-05-07
> I guess we should remove object.di from the source folder now?

Yep we should, I'd still like to copy object.d instead of renaming it to .di.
For a certain period we might add an `rm -f import/object.di` to avoid using a stall header file instead of `import/object.d`.